requirements.txt
boto3==1.9.6
botocore==1.12.6
Django==2.1.1
Django-Accounts==0.1
django-crispy-forms==1.7.2
django-geoposition==0.3.0
djangorestframework==3.8.2
docutils==0.14
jmespath==0.9.3
Pillow==5.2.0
python-dateutil==2.7.3
pytz==2018.5
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
